requirements.txt
asgiref==3.5.2
Django==3.2.5
django-crispy-forms==1.14.0
django-jazzmin==2.5.0
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
